# Inline symbols