--- a/compiler/vm/src/heap/object_heap/tag.rs
+++ b/compiler/vm/src/heap/object_heap/tag.rs
@@ -28,7 +28,7 @@ impl HeapTag {
         heap: &mut Heap,
         is_reference_counted: bool,
         symbol_id: SymbolId,
-        value: impl Into<Option<InlineObject>>,
+        value: impl Into<InlineObject>,
     ) -> Self {
         let symbol_id = symbol_id.value();
         debug_assert_eq!(
@@ -45,7 +45,7 @@ impl HeapTag {
             2 * HeapObject::WORD_SIZE,
         ));
         unsafe {
-            *tag.value_pointer().as_mut() = value.map_or(0, |value| value.raw_word().get());
+            *tag.value_pointer().as_mut() = value.raw_word().get();
         };
         tag
     }
@@ -58,12 +58,9 @@ impl HeapTag {
     fn value_pointer(self) -> NonNull<u64> {
         self.content_word_pointer(0)
     }
-    pub fn has_value(self) -> bool {
-        unsafe { *self.value_pointer().as_ref() != 0 }
-    }
-    pub fn value(self) -> Option<InlineObject> {
+    pub fn value(self) -> InlineObject {
         let value = unsafe { *self.value_pointer().as_ref() };
-        NonZeroU64::new(value).map(InlineObject::new)
+        InlineObject::new(NonZeroU64::new(value).unwrap())
     }
 }
 
@@ -71,10 +68,7 @@ impl Debug for HeapTag {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         write!(f, "<symbol-id {}>", self.symbol_id().value())?;
 
-        if let Some(value) = self.value() {
-            write!(f, " ({value:?})")?;
-        }
-        Ok(())
+        write!(f, " ({:?})", self.value())
     }
 }
 impl DisplayWithSymbolTable for HeapTag {
@@ -82,14 +76,11 @@ impl DisplayWithSymbolTable for HeapTag {
         // We can always use the display formatter since the symbol has a constrained charset.
         write!(f, "{}", symbol_table.get(self.symbol_id()))?;
 
-        if let Some(value) = self.value() {
-            write!(
-                f,
-                " ({})",
-                DisplayWithSymbolTable::to_string(&value, symbol_table),
-            )?;
-        }
-        Ok(())
+        write!(
+            f,
+            " ({})",
+            DisplayWithSymbolTable::to_string(&self.value(), symbol_table),
+        )
     }
 }
 
@@ -103,9 +94,7 @@ impl PartialEq for HeapTag {
 impl Hash for HeapTag {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.symbol_id().hash(state);
-        if let Some(value) = self.value() {
-            value.hash(state);
-        }
+        self.value().hash(state);
     }
 }
 
@@ -131,19 +120,15 @@ impl HeapObjectTrait for HeapTag {
         clone: HeapObject,
         address_map: &mut FxHashMap<HeapObject, HeapObject>,
     ) {
-        let value = self
-            .value()
-            .map(|it| it.clone_to_heap_with_mapping(heap, address_map));
+        let value = self.value().clone_to_heap_with_mapping(heap, address_map);
         let clone = Self(clone);
         unsafe {
-            *clone.value_pointer().as_mut() = value.map_or(0, |it| it.raw_word().get());
+            *clone.value_pointer().as_mut() = value.raw_word().get();
         };
     }
 
     fn drop_children(self, heap: &mut Heap) {
-        if let Some(value) = self.value() {
-            value.drop(heap);
-        }
+        self.value().drop(heap);
     }
 
     fn deallocate_external_stuff(self) {}