… compiler/vm/src/heap/object_heap/text.rs
… compiler/vm/src/heap/object_inline/int.rs
… compiler/vm/src/heap/object_inline/mod.rs
… compiler/vm/src/heap/object_inline/port.rs
… compiler/vm/src/heap/object_inline/tag.rs
… compiler/vm/src/heap/object_representation.md
… compiler/vm/src/heap/symbol_table.rs
… compiler/vm/src/lib.rs
… compiler/vm/src/mir_to_lir.rs
… compiler/vm/src/tracer/stack_trace.rs
… compiler/vm/src/vm.rs