**This depends on #590!**

<!-- Please enter the corresponding issue ID: -->


<!-- Please summarize your changes: -->
Tags without a value are now represented as inline values, avoiding allocations.

## Benchmark Results

`fibonacci 20` instead of the default `fibonacci 15`

```
Before: cargo run --release -- run packages/examples/fibonacci.candy
  Time (mean ± σ):     932.2 ms ±  82.4 ms    [User: 893.5 ms, System: 38.8 ms]
  Range (min … max):   857.2 ms … 1128.0 ms    10 runs

After: cargo run --release -- run packages/examples/fibonacci.candy
  Time (mean ± σ):     887.8 ms ±  49.3 ms    [User: 850.9 ms, System: 37.3 ms]
  Range (min … max):   815.4 ms … 943.7 ms    10 runs
```

(Before is #590, not main)

### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)